reorganize and lint spec folder
