+++ spec/bic_candidate_spec.rb
@@ -1,4 +1,6 @@
 RSpec.describe IbanCalculator::BicCandidate do
+  include_context 'bic_candidate'
+
   subject { described_class.new(single_candidate[:item]) }
 
   describe '.build_list' do
@@ -63,42 +65,5 @@ RSpec.describe IbanCalculator::BicCandidate do
         'www_count' => 0
       )
     end
-  end
-
-  def single_candidate
-    {
-      :item => {
-        :bic => 'BOFIIE2D',
-        :zip => 'zip',
-        :city => 'city',
-        :wwwcount => '0',
-        :sampleurl => 'sample_url',
-        :'@xsi:type' => 'tns:BICStruct'
-      },
-      :'@xsi:type' => 'SOAP-ENC:Array',
-      :'@soap_enc:array_type' => 'tns:BICStruct[1]'
-    }
-  end
-
-  def multiple_candidates
-    {
-      :item => [{
-        :bic => 'BOFIIE2D',
-        :zip => 'zip',
-        :city => 'city',
-        :wwwcount => '0',
-        :sampleurl => 'sample_url',
-        :'@xsi:type' => 'tns:BICStruct'
-      }, {
-        :bic => 'BOFIIE2D',
-        :zip => 'zip',
-        :city => 'city',
-        :wwwcount => '0',
-        :sampleurl => 'sample_url',
-        :'@xsi:type' => 'tns:BICStruct'
-      }],
-      :'@xsi:type' => 'SOAP-ENC:Array',
-      :'@soap_enc:array_type' => 'tns:BICStruct[1]'
-    }
   end
 end
